# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: BookCon API
  version: "1.0"
  description: Self-hosted eBook library & sync API. Auth via Bearer JWT (15 min) + rotating refresh tokens. Deletions are tombstones; every syncable entity carries updated_at/deleted_at.
servers:
  - url: "{base}/api/v1"
    variables:
      base:
        default: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    limitParam:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 500 }
  schemas:
...
    BookBase:
      type: object
      required: [title, format]
      properties:
        title: { type: string }
        authors: { type: array, items: { type: string }, default: [] }
        description: { type: string, default: "" }
        language: { type: string, nullable: true }
        publisher: { type: string, nullable: true }
        published_date: { type: string, nullable: true }
        series_id: { type: string, format: uuid, nullable: true }
        series_index: { type: number, nullable: true }
    Book:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            id: { type: string, format: uuid }
            format: { type: string, enum: [epub, pdf, cbz, cbr] }
            status: { type: string, enum: [processing, ready, failed], default: ready }
            status_message: { type: string, nullable: true }
            cover_url: { type: string, nullable: true }
            file_size_bytes: { type: integer, nullable: true }
            page_count: { type: integer, nullable: true }
            word_count: { type: integer, nullable: true }
            tag_ids: { type: array, items: { type: string, format: uuid } }
            shelf_ids: { type: array, items: { type: string, format: uuid } }
            added_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            deleted_at: { type: string, format: date-time, nullable: true }
...
    InitiateUploadRequest:
      type: object
      required: [filename, size_bytes, sha256, content_type]
      properties:
        filename: { type: string }
        size_bytes: { type: integer }
        sha256: { type: string, minLength: 64, maxLength: 64 }
        content_type: { type: string }
    InitiateUploadResponse:
      oneOf:
        - type: object
          required: [outcome, upload_url, book_id]
          properties:
            outcome: { type: string, const: upload_required }
            upload_url: { type: string }
            method: { type: string, const: PUT }
            expires_in: { type: integer }
        - type: object
          required: [outcome, book_id]
          properties:
            outcome: { type: string, const: duplicate }
            book_id: { type: string, format: uuid }
...
paths:
...
  /books:
    get:
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - $ref: "#/components/parameters/limitParam"
        - { name: q, in: query, schema: { type: string } }
        - { name: shelf_id, in: query, schema: { type: string, format: uuid } }
        - { name: tag_id, in: query, schema: { type: string, format: uuid } }
        - { name: series_id, in: query, schema: { type: string, format: uuid } }
        - { name: format, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [recent, added, title, author, progress], default: recent } }
      responses:
        "200":
          description: Page of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Book" } }
                  next_cursor: { type: string, nullable: true }
  /books/initiate-upload:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InitiateUploadRequest" }
      responses:
        "201": { description: Presigned URL or dedupe hit, content: { application/json: { schema: { $ref: "#/components/schemas/InitiateUploadResponse" } } } }
  /books/{book_id}:
    get:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Book detail, content: { application/json: { schema: { $ref: "#/components/schemas/Book" } } } }
        "404": { description: Not found }
    patch:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/BookBase"
                - type: object
                  properties:
                    tag_ids: { type: array, items: { type: string, format: uuid } }
                    shelf_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        "200": { description: Updated book }
    delete:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Tombstoned }
  /books/{book_id}/complete-upload:
    post:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "202": { description: Processing started, content: { application/json: { schema: { $ref: "#/components/schemas/Book" } } } }
  /books/{book_id}/file-url:
    get:
      parameters: [{ name: book_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: Presigned download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  expires_in: { type: integer }
...